# apigateway/openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce API Gateway
  version: 1.1.0
  description: Consolidated OpenAPI spec for the API Gateway proxying User, Product, and Order services.
servers:
  - url: http://localhost:8083
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      description: Prevent duplicate order creation. If provided, repeated requests return the same order.
      schema:
        type: string
        maxLength: 64
  schemas:
    Error:
      type: object
      properties:
        message: { type: string }
        code: { type: string }
        details: { type: object, additionalProperties: true }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string, format: email }
        token: { type: string, description: Bearer JWT token }
    Address:
      type: object
      properties:
        id: { type: string }
        line1: { type: string }
        line2: { type: string, nullable: true }
        city: { type: string }
        state: { type: string }
        postal_code: { type: string }
        country: { type: string }
        phone: { type: string, nullable: true }
        is_default: { type: boolean }
...
    User:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
    CreateUserRequest:
      type: object
      required: [username, email, password]
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password }
        phone: { type: string }
    Product:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer }
    CreateProductRequest:
      type: object
      required: [name, price, stock]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
    UpdateProductRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, format: float }
        stock: { type: integer, minimum: 0 }
...
    OrderStatus:
      type: string
      enum: [PENDING, PAID, CANCELLED]
    OrderItem:
      type: object
      properties:
        product_id: { type: string }
        quantity: { type: integer }
        price: { type: number }
    Order:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        total_amount: { type: number }
        shipping_address_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    CreateOrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }
    CreateOrderRequest:
      type: object
      required: [userId, items]
      properties:
        userId: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItem'
        shippingAddressId: { type: string, nullable: true }
    OrdersListResponse:
      type: object
      properties:
        orders:
          type: array
          items: { $ref: '#/components/schemas/Order' }
        nextCursor: { type: string, nullable: true }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # Users
  /api/v1/users:
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/users/{userId}:
    get:
      tags: [Users]
      summary: Get user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateProductRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
  /api/v1/products/{productId}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    put:
      tags: [Products]
      summary: Update product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProductRequest' }
      responses:
        '200': { description: Updated }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Deleted }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  status: { $ref: '#/components/schemas/OrderStatus' }
        '200': { description: Returned existing order for idempotent request }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '503': { description: Dependency unavailable }
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: userId
          schema: { type: string }
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrdersListResponse' }
  /api/v1/orders/{orderId}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...